Minimal patch to fix #2290 (/tmp permmisions)
This patch rectifies the specified problem by resetting the
permissions to 1777 after the offending tests have run.  I
am providing it for testing but suspect there may be a deeper
bug hiding behind this one--specifically, it isn't clear to
me that an analogous problem could not occur in production use
(which this patch to the spec would not fix).

Signed-off-by: Markus Roberts <Markus@reality.com>