Interim fix for #2615 (templates picking up extra "/n"s)
This reduces the cases of #2615 to the level that were present in
0.24.x by removing the optional (and problematic) domain-boundary
"\n"s in the generated yaml.  This means the only times serialized
text values should pick up an additional trailing "\n" is if they
contain two or more trailing "\n" to start with and either occur
at the very end of the stream or directly before a Regex or DateTime
used as a Hash key (i.e. we should almost never see it).

Signed-off-by: Markus Roberts <Markus@reality.com>