Fix for #2731 (problem communicating with processes in SELinux)
In some cases communicating with child processes via temprary files is not
viable.  This is Ricky Zhou's patch from the ticket, which solves the problem
by using the more normal system of pipes.  It is a broader reaching change
than suggested by the ticket (it affects all execs, not just SELinux) but IMHO
is the right way to go.

Signed-off-by: Markus Roberts <Markus@reality.com>