Proposed fix for #3013 (pipes, OSes, and ruby versions, Oh My!)
The recent release of 0.25.2 contains a new method of handling communication
with "exec"ed processes to support SELinux and improve performance on other
OSes which, unfortunately, has a number of shortcomings:

* It does not work on ruby < 1.8.3
* It reintroduces one/possibly two bugs
* It has no tests

This patch addresses these problems (and an analogous issue with SELinux mount
handling discovered in the process).  The structure of the patch:

* Tests for the base cases and applicable bugs from the ticket history
* Code to select between readpartial (on ruby >=1.8.3) and sysread (<1.8.3)
* Logic to detect and deal with first and second order failures/misbehavior
  in the "exec"ed process
* Logic to handle asynchronous signals (also ported to the SELinux mount code)
* Notes / tests (pending, but written) for the remaining, unhandled third-
  order failure case

This last bears some elaboration; if the following steps occur:

* Puppet "exec"s child process (apt-get, for example) and attempts to capture
  the output
* The child process launches a "grandchild" process (the buggy version of the
  osirid daemon under Debian Etch, for example) without closing it's STDOUT
* The grandchild does not close its inherited STDOUT
* The child exits
* The grandchild does not trap SIGPIPE
* The grandchild writes to STDOUT

the grandchild will receive an (unhandled) SIGPIPE and exit.  There is no
presently known case where this would occur (the osirid daemon, as best as I
can determine, did not actually write to the handle, just kept it around out
of misplaced sentimentality).  The solutions to this problem, if it exists,
are messy and best left unimplemented unless they are needed.