Refinement of fix for #3013
This patch incorporates Ricky Zhou's idea for establishing a default SIGPIPE
trap of IGNORE in child, which protects really dumb daemons from the effects of
writing to a closed stdout.  As even moderately smart daemons ought not do
this, these may be a reasonable fraction of the cases.

It also includes a sketch of the test for the patch, which is marked pending
as there is no good way to implement it as a portable automated test; you
either need a small standalone C program, a hacked ruby interpreter, or other
shenanigans to run it.

Also, tightened tests (~40 seconds instead of ~200).