Re-adding a read_cert equivalent.
Some of the code that wasn't delete use a routine (read_cert, which
read in the certificate from "#{Puppet[:hostname]}.pem" with some
logic to deal with mixed-case name) which was dynamically generated
and added to various classes by the code that was deleted.

As there was no place in the undeleted code where this was done, this
patch re-adds that functionality but without all the metaprogramming
flourishes.