Fix for #3107 Changing users on AIX
There are several issues with changing the real, effective, and saved group
and user ids in different environments (which methods to call, in what order,
etc).  While the code being replaced by this patch appeared to work for Linux,
Solaris, and (with a special case test) Darwin; it was failing under AIX and
may have had edge-case problems under the others.

Ruby back to 1.8.1 has supported a higher level interface that deals with the
problem and captures a broader range of OSes; it's a single call for group and
one for user--the details of rid/eid/svid, etc ordering are handled internally.
Switching to that simplifies our code and should improve/unify our support of
various OSes.