Improving fix for #1175; tightening thread safety
The previous code maintained thread safety up to work-duplication (so that a
collision would, at worse, result in effective cache flushing and cause some
additional work to be done).  The preceding patch addressed the single thread
issue of environment specific functions; this patch brings the thread safety
up to the previous standard.