Alt fix for #4207 -- serialize environments as their names
Environments contain a deal of transitory information and references to other
objects, none of which is wanted when they are serialized.

Rather than having this serialization concern propogate through the code by
replacing environments by their names prior to serialization (which would be
one way to address the problem) this patch changes environments so they only
serialize their identity (name) and not their contents.