Fix #4302 - Compilation speed regression compared to 2.6
Each time the compiler was accessing the loaded types, we were checking if the
manifests had changed.  This incurred a large performance cost compared to 0.25
and introduced race conditions if manifests changed while a thread was in the
middle of a compilation.

This tentative fix, based on Brice's, makes sure each thread will get access to
the same loaded types collection for the durration of a compilation, even if
the manifests change.  We now only check for changed files at the start of a
compilation or if the environment changes, and we maintain a per environment
thread lock so that only one thread at a time can be reloading any particular
environment (and the need-check is done inside the synchronize block so that
only the first will actually load it).

As long as the manifests don't change, the threads will share the same collection,
so there is only duplication in memory for a brief window surrounding a change.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>
Second-author: Markus Roberts <markus@puppetlabs.com>