Possible fix for #4297
This fix is based on refactoring Jesse's empirically derived solution guided by
the assumption that #4297 was introduced / exposed by the fix for #4270; I do
not have a global understanding of the code here, but am just making an
adjustment that 1) slightly reduces the code complexity, 2) doesn't appear to
break anything, 3) empirically fixes the observed problem, 4) makes since in
that resource classes shouldn't have to consult a passed-in scope to find their
own namespaces.

Basically, I'm fixing what I believe was a thinko in a patch that I don't fully
understand.