Fixes #4485 -- single quoted strings should not treat \n as new line
This is the rest of the change for #4303; James and I discussed various
ways the solution to that ticket needed to be extended but, as neither of us
committed code, nothing changed.  This is the least implact extension, which
mimics the behaviour of prior versions.

It leaves open the question: should '\\x' start with a single or double
backslash?  If, as now, '\\x' starts with a double backslash (i.e. single quote
is the only escapable characterin single quoted strings) a string ending in a
backslash can not be represented in a single quoted string.