Maint. -- Fix test failures broken by previous commit
This basically involved adding a search method to the yaml indirector, which I
did by copying the one from ssl_file and fiddling with it until the tests
passed.  Since the most straight forward way to do this required extending the
interface to the path method I added tests for the additional behaviour.