Step towards #5027 -- add Logging#deprication_warning facility
This commit adds a method analogous to Puppet.warn which 1) only logs each
message the first time it is received and 2) only logs the first 100 messages
it receives.  Messages are logged via warn.

This could easily be made more flexible by making the hard limit and effective
log level user settable, if desired.