Fix for #5298 -- Collections need to do type lookup
When the responsibility for type-name resolution was moved to the AST nodes in
commit 449315a2c705df2396852462a1d1e14774b9f117, at least one instance was
missed: the space ship operator

    Myclass <<| tag == foo |>>

fails unless Myclass has been previously loaded.  This commit adds the lookup
to AST::Collection nodes in the same way it was added to the other node types.

Note that I haven't audited the other note types for similar cases.